requirements.txt
flask
cmake
setuptools
wheel
ipykernel
ipywidgets
langchain
PyMuPDF
chromadb
sentence-transformers
llama-cpp-python
openai
langchain-community
langchain-huggingface
sentencepiece
llama-stack
fairscale
streamlit
